data-engineering/bitinfocharts_ws/requirements.txt
# python library dependencies
beautifulsoup4
python-dotenv
sqlalchemy
psycopg2
pandas
mypy
lxml